+++ tests/testfreq.c
@@ -1,7 +1,7 @@
 
 /* 
  * Very simple test program to check freq convertion --SF
- * This is mainly to test kHz, MHz, GHz macros and long long support.
+ * This is mainly to test kHz, MHz, GHz macros and int64_t support.
  */
 
 #ifdef HAVE_CONFIG_H
@@ -36,27 +36,27 @@ int main (int argc, char *argv[])
 
 	/* freq on 31bits test */
 	f = GHz(2);
-	printf("GHz(2) = %"PRIll"\n", (long long)f);
+	printf("GHz(2) = %"PRIll"\n", (int64_t)f);
 
 	/* freq on 32bits test */
 	f = GHz(4);
-	printf("GHz(4) = %"PRIll"\n", (long long)f);
+	printf("GHz(4) = %"PRIll"\n", (int64_t)f);
 
 	/* freq on >32bits test */
 	f = GHz(5);
-	printf("GHz(5) = %"PRIll"\n", (long long)f);
+	printf("GHz(5) = %"PRIll"\n", (int64_t)f);
 
 	/* floating point to freq conversion test */
 	f = GHz(1.3);
-	printf("GHz(1.3) = %"PRIll"\n", (long long)f);
+	printf("GHz(1.3) = %"PRIll"\n", (int64_t)f);
 
 	/* floating point to freq conversion precision test */
 	f = GHz(1.234567890);
-	printf("GHz(1.234567890) = %"PRIll"\n", (long long)f);
+	printf("GHz(1.234567890) = %"PRIll"\n", (int64_t)f);
 
 	/* floating point to freq conversion precision test, with freq >32bits */
 	f = GHz(123.456789012);
-	printf("GHz(123.456789012) = %"PRIll"\n", (long long)f);
+	printf("GHz(123.456789012) = %"PRIll"\n", (int64_t)f);
 
 	return 0;
 }
